requirements.txt
fastapi
uvicorn
Pillow
face_recognition
torch
transformers
opencv-python
numpy